タスクカードを追加時、最新のものが画面外に表示される #21
